+++ Runtime/Serializers/ISerializer.cs
@@ -5,6 +5,7 @@ namespace ActionCode.Persistence
         string Extension { get; }
 
         string Serialize<T>(T data);
+        string SerializePretty<T>(T data);
         T Deserialize<T>(string value);
     }
 }