+++ Runtime/Serializers/NewtonsoftSerializer.cs
@@ -9,9 +9,11 @@ namespace ActionCode.Persistence
         public string Extension => "json";
 
 #if PACKAGE_NEWTONSOFT
+        public string SerializePretty<T>(T data) => JsonConvert.SerializeObject(data, Formatting.Indented);
         public string Serialize<T>(T data) => JsonConvert.SerializeObject(data);
         public T Deserialize<T>(string value) => JsonConvert.DeserializeObject<T>(value);
 #else
+        public string SerializePretty<T>(T data) => Serialize(data);
         public string Serialize<T>(T _) => string.Empty;
         public T Deserialize<T>(string _) => default;
 #endif